requirements.txt
streamlit
pytesseract
Pillow
pdf2image
requests
opencv-python
numpy